+++ api/models/__init__.py
@@ -1,2 +1,3 @@
 from brew_pi_spark import BrewPiSpark
+from configuration import Configuration
 from device import Device