+++ api/tasks/__init__.py
@@ -1,1 +1,2 @@
 from status_message import check_if_status_update_required
+from logs_message import log_device_data