+++ api/views/spark_status.py
@@ -12,12 +12,11 @@ logger = logging.getLogger(__name__)
 
 
 @require_http_methods(["POST"])
-@csrf_exempt
 def check_in(request):
     logger.info("Check in request received")
 
     spark = BrewPiSpark.from_json(request.body)
 
-    status_message.check_if_status_update_required.delay(spark.device_id)
+    status_message.check_if_status_update_required.delay(spark.device_id, request.META['SERVER_PORT'])
 
     return ApiResponse.ok()