+++ celery_dev_server.sh
@@ -1,0 +1,1 @@
+celery -A oinkbrew_webapp worker -l info